requirements.txt
attrs==23.1.0
boolean.py==4.0
build==1.0.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cyclonedx-python-lib @ git+ssh://git@github.com/pre-quantum-research/cyclonedx-python-lib.git@cbom#egg=cyclonedx-python-lib
defusedxml==0.7.1
gitchangelog==3.0.4
idna==3.4
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
license-expression==30.1.1
packageurl-python==0.11.2
packaging==23.2
pbr==6.0.0
pluggy==1.3.0
py-serializable==0.11.1
pycparser==2.21
pyproject_hooks==1.0.0
pystache==0.6.5
pytest==7.4.3
pyyaml==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
sortedcontainers==2.4.0
urllib3==2.0.7
